refactor: move grid and state helpers to modules
